// src/app/models/wall-post.model.ts
export class WallPost {
    title!: string;
    description!: string;
    createdDate!: Date;
    imageUrl!: string;
    location?: string ;
    author?: string;
    _id?:string;
    likes?: number;
    createdAt?:Date;
    updatedAt?: Date;
    nameUser?: string;
    prenomUser?: String;
    // id?:number;
}